Implemented the client connection to the server (#1)
Initial implementation of reading standard input to initialize the server list. It also tries to connect to the closest server once the connection is lost or first stablished. It gives 100ms to try to connect to a server, after that it jumps to the next closest server looping until a connection can be stablished.
If a message is not send connection is reset. If connection is ended while waiting for the response of the server, reconnection process starts and sends the message again.
Server side must be responsible of handling this.